feat: implement comprehensive currency management system with integer storage
- Add currency configuration to BusinessInertiaShareProvider for sharing via props
- Update all database migrations to use integer storage for monetary values
- Convert all Model casts from decimal/float/MoneyCast to integer for price fields
- Update all DTOs to use integers instead of floats for price properties
- Create reusable CurrencyInput component with locale-aware formatting
- Add parseCurrencyInput() and formatCurrencyForInput() utility functions
- Update all frontend forms to use new CurrencyInput component
- Fix OrderCalculationService to work with integer arithmetic
- Remove MoneyCast dependencies that caused type mismatch errors

All monetary values are now stored as integers in the smallest currency unit
(cents for USD, fils for KWD, whole units for CLP) to avoid floating-point
precision issues. Currency configuration comes from the database and is
automatically shared with frontend via Inertia props.

+++ app-modules/business/database/migrations/2025_08_27_171726_create_business_subscriptions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             // Plan information
             $table->string('plan_id'); // basic, pro, enterprise
             $table->string('plan_name');
-            $table->decimal('price', 10, 2);
+            $table->integer('price'); // Stored in minor units (cents, fils, etc.)
             $table->string('currency', 3)->default('CLP');
             $table->enum('billing_cycle', ['monthly', 'yearly', 'lifetime'])->default('monthly');
