+++ app-modules/item/database/migrations/2025_07_29_000004_create_item_modifiers_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('modifier_group_id')->constrained()->cascadeOnDelete();
             $table->string('name');
-            $table->decimal('price_adjustment', 10, 2)->default(0); // Can be positive or negative
+            $table->integer('price_adjustment')->default(0); // Can be positive or negative, stored in minor units
             $table->integer('max_quantity')->default(1);
             $table->boolean('is_default')->default(false);
             $table->boolean('is_active')->default(true);
